+++ th2_data_services/data.py
@@ -13,6 +13,7 @@
 #  limitations under the License.
 import copy
 import csv
+import json
 import gc
 import pickle
 import pprint
@@ -920,6 +921,28 @@ class Data(Generic[DataIterValues]):
 
         return self
 
+
+    def to_json(self, filename: str, indent: int = None, overwrite: bool = False):
+        """Converts data to json format.
+
+        Args:
+            filename (str): Output JSON filename
+            indent (int, optional): JSON format indent. Defaults to None.
+            overwrite (bool, optional): Overwrite if filename exists. Defaults to False.
+
+        Raises:
+            FileExistsError: If file exists and overwrite=False
+        """
+        if Path(filename).absolute().exists() and not overwrite:
+            raise FileExistsError(f"{filename} already exists. If you want to overwrite current file set `overwrite=True`")
+
+        with open(filename, 'w', encoding='UTF-8') as file:
+            file.write('[') # Start list
+            for record in self:
+                json.dump(record, file, indent=indent)
+                file.write(',\n')
+            file.seek(file.tell() - 3) # Delete last comma for valid JSON
+            file.write(']') # Close list
 
 def _iter_any_file(filename, mode="r"):
     """Returns the function that returns generators."""