+++ release_notes.md
@@ -473,6 +473,7 @@ If you want to use RDP you have to specify dependency in square brackets `[ ]`
 15. [TH2-4881] Data.from_json method was added.
 16. [TH2-4919] Data.from_any_file method was added.
 17. [TH2-4928] Data.from_csv method was added.
+18. [TH2-4932] Data.to_json method was added.
 
 ## BugFixes
 1. [TH2-4711] EventTreeCollection max_count parameter of findall functions worked wrong.